Major changes
usage for RemoveDuplicates.
Solve Quietly....

ReZAnd[a,b][c_And] probably not needed.

Simplify Or in NewReduce.

Other definitions in NewReduce

Simplified EqEliminator, the heart of CleanEqualities.

CriticalBundle now returns the d2e too.

CriticalCongestionSolver now solves for u variables before the fixed point (the fixed point starts with NewReduce, so we kill as much ORs as we can)