Improved D2E
Switching costs compatibility is simpler (and faster).

put some stuff together so it simplifies the reading.

we are solving or giving solutions to all the equations in the system.

cost equations on the vertices! easier to extract equalities when the switching cost is zero.

returning substituted inequalities, ors, and criticalcongestion equations.

critical congestion rules too.