data2equations is better
we can still make it better......
(remove some jt expressions,
construct some jts as rules instead of equations in EqBalanceSplitting/.RulesBalanceGathering)

removed some commented code.

we are giving messages to the user.

using clean equalities!!

we are not simplifying big chunks of the system!

returning RulesCriticaCase, EqPosJs, EqCurrentCompCon.

This way we don't have to reduce the whole system. we get the values for us and substitute back in the original EqCriticalCase with the conditions on js.