Update D2E2.m
GetKirchhoff works. We could also ask for more.

We are replacing new rules correctly in the joined rules, now.

Changed compositional application of TripleClean for the FinalClean function.

Picking one solution of the system when there are (possibly) multiple solutions.
This helps with determining all currents (even transition ones!)

Introduced the functions FinalStep and FinalClean.